+++ script/script.js
@@ -2,8 +2,8 @@
 import { questoesHtml } from "./temaHtml.js";
 import { questoesCSS } from "./temaCSS.js";
 import { questoesJS } from "./temaJS.js";
-import {pararCronometro} from "./cronometro.js";
-import {iniciarCronometro}  from "./cronometro.js"
+import {pararCronometro, iniciarCronometro} from "./cronometro.js";
+import { comVolume, semVolume, btnAudio, audioSvg } from "./trilhaSonora.js";
 
 //========= Declarações =============================
 
@@ -326,35 +326,8 @@ function ranking() {
 mostrarTela(home);
 
 //========= Trilha Sonora =====================
-
-// achando os elementos no HTML
-
-const audio = document.getElementById("audio");
-audio.volume = 0.1;
-const btnAudio = document.getElementById("btn-audio");
-let audioSvg;
 let stopped = true;
 
-// função para dar volume no som
-function comVolume() {
-    audio.volume = 0.1;
-    //atribui o ícone de "com audio" quando a função for chamada
-    audioSvg = `<svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-volume-up-fill" viewBox="0 0 16 16">
-    <path d="M11.536 14.01A8.473 8.473 0 0 0 14.026 8a8.473 8.473 0 0 0-2.49-6.01l-.708.707A7.476 7.476 0 0 1 13.025 8c0 2.071-.84 3.946-2.197 5.303l.708.707z"/>
-    <path d="M10.121 12.596A6.48 6.48 0 0 0 12.025 8a6.48 6.48 0 0 0-1.904-4.596l-.707.707A5.483 5.483 0 0 1 11.025 8a5.483 5.483 0 0 1-1.61 3.89l.706.706z"/>
-    <path d="M8.707 11.182A4.486 4.486 0 0 0 10.025 8a4.486 4.486 0 0 0-1.318-3.182L8 5.525A3.489 3.489 0 0 1 9.025 8 3.49 3.49 0 0 1 8 10.475l.707.707zM6.717 3.55A.5.5 0 0 1 7 4v8a.5.5 0 0 1-.812.39L3.825 10.5H1.5A.5.5 0 0 1 1 10V6a.5.5 0 0 1 .5-.5h2.325l2.363-1.89a.5.5 0 0 1 .529-.06z"/>
-    </svg>`;
-}
-
-//função para deixar o som sem volume
-function semVolume() {
-    audio.volume = 0;
-    //atribui o ícone de "sem som" quando a função for chamada
-    audioSvg = `<svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-volume-mute-fill" viewBox="0 0 16 16">
-    <path d="M6.717 3.55A.5.5 0 0 1 7 4v8a.5.5 0 0 1-.812.39L3.825 10.5H1.5A.5.5 0 0 1 1 10V6a.5.5 0 0 1 .5-.5h2.325l2.363-1.89a.5.5 0 0 1 .529-.06zm7.137 2.096a.5.5 0 0 1 0 .708L12.207 8l1.647 1.646a.5.5 0 0 1-.708.708L11.5 8.707l-1.646 1.647a.5.5 0 0 1-.708-.708L10.793 8 9.146 6.354a.5.5 0 1 1 .708-.708L11.5 7.293l1.646-1.647a.5.5 0 0 1 .708 0z"/>
-    </svg>`;
-}
-// stopped inicia como verdade então o primeiro clique vai entrar no if, dentro do if o valor de stopped vai mudar e no segundo clique ele vai entrar no else, e assim sucessivamente.
 btnAudio.onclick = () => {
     if(stopped) {
         console.log("teste if", stopped)
@@ -366,7 +339,7 @@ btnAudio.onclick = () => {
         semVolume();
         stopped = true;
     }
-    // troca o ícone toda vez que clicamos no botão seguindo as funções play/pause
+    // troca o ícone toda vez que clicamos no botão seguindo as funções comVolume/semVolume
     btnAudio.innerHTML = audioSvg;  
 }
